Adds cli tool for managing stac schemas
methods for listing, getting, validating schemas.
concrete implementation of compiled schema registry, meaning that
schemas are defined via go source files at build time. tools are
provided to render a self-contained json-schema document from a
json-schema fragment in yaml which represents the properties ssection of
a geo-json feature. this is intended to be more convenient than working
with the full boilerplate geo-json schema, as the properties section is
probably the only part of the geo-json document schema that should be
extended by stac implementations.